requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib
bcrypt==4.0.1
python-jose[cryptography]
python-dotenv
pytesseract
Pillow
python-multipart
google-generativeai
requests
